Tipizition full

+++ src/components/selectWorkoutBtn/index.tsx
@@ -5,9 +5,11 @@ const SelectWorkoutBtn = ({ id }: {id: string}) => {
     const router = useRouter();
 
   const goToTheWorkoutPage = () => {
-    const localIdWorkout = localStorage.getItem('idWorkout');
-    router.push(`/workout-page/${id}/${localIdWorkout}`);
-    localStorage.clear();
+      const localIdWorkout = localStorage.getItem('idWorkout');
+      if (localIdWorkout) {
+        router.push(`/workout-page/${id}/${localIdWorkout}`);
+        localStorage.clear();
+      } 
   }
   return (
     <button className="w-full bg-bright-green rounded py-4 px-7 text-sm text-black mt-[34px]" onClick={goToTheWorkoutPage}>Начать</button>
